fix: link to delivery info, order card product price
