+++ components/Order/Card/Card.styles.ts
@@ -1,0 +1,45 @@
+const sx = {
+  card: {
+    width: '100%',
+    padding: '30px',
+    backgroundColor: 'background.paper',
+    borderRadius: '6px',
+  },
+  footer: {
+    width: '100%',
+    borderTop: '1px dashed',
+    borderColor: 'text.muted',
+    margin: '20px 0 0 0',
+    padding: '20px 0 0 0',
+  },
+  field: {
+    display: 'flex',
+    flexDirection: {
+      xs: 'column',
+      sm: 'row',
+      md: 'column',
+    },
+    gap: {
+      xs: 0,
+      sm: '20px',
+      md: 0,
+    },
+    marginBottom: '10px',
+  },
+  count: {
+    fontFamily: 'Roboto slab',
+    fontWeight: 'bold',
+    margin: '0 0 20px 0',
+  },
+  product: {
+    width: 'fit-content',
+    color: 'text.muted',
+  },
+  total: {
+    fontFamily: 'Roboto slab',
+    fontWeight: 'bold',
+    color: 'text.secondary',
+  },
+};
+
+export default sx;
