+++ pages/basket/Basket.styles.ts
@@ -1,0 +1,25 @@
+const sx = {
+  title: {
+    fontSize: {
+      sm: '40px',
+      xs: '24px',
+    },
+    fontFamily: 'Roboto slab',
+    fontWeight: 'bold',
+    color: 'text.secondary',
+  },
+  divider: {
+    borderColor: 'secondary.main',
+  },
+  desktopOrderBtn: {
+    display: { xs: 'none', md: 'flex' },
+    width: '100%',
+    marginBottom: '10px',
+  },
+  mobileOrderBtn: {
+    width: '100%',
+    marginTop: '10px',
+  },
+};
+
+export default sx;
